factor out Workspace to workspace.py
