+++ .github/workflows/ci.yml
@@ -45,7 +45,7 @@ jobs:
       if: matrix.python-version == 3.7
       # FIXME: numpy_minuit-200_bins excluded due to speed regression from PR #553. c.f. Issue #572
       run: |
-        python -m pytest -r sx --benchmark-sort=mean tests/benchmarks/test_benchmark.py -k 'not numpy_minuit-200_bins'
+        python -m pytest -r sx --benchmark-sort=mean tests/benchmarks/test_benchmark.py
 
   docs:
