+++ src/pyhf/__init__.py
@@ -81,7 +81,8 @@ def set_backend(backend, custom_optimizer=None):
         events.trigger("optimizer_changed")()
 
 
-from .pdf import Model, Workspace
+from .pdf import Model
+from .workspace import Workspace
 from . import simplemodels
 
 __all__ = ['Model', 'Workspace', 'utils', 'modifiers', 'simplemodels', '__version__']
